pyproject.toml
[project]
name = "fabric-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
]
